+++ src/ts/ajax.ts
@@ -39,16 +39,15 @@ export class Ajax {
       http.onreadystatechange = () => {
         if(http.readyState == 4) {
           if(http.status == 200) {
-            console.log("xhr done successfully");
+            console.log('xhr done successfully ('+url+')');
             resolve(http.responseText);
           } else {
-            console.log("xhr failed");
+            console.log('error', 'xhr failed ('+url+')');
             reject(http.status);
           }
-        } else {
-          console.log("xhr processing going on");
         }
       };
+      console.log('xhr processing starting ('+url+')');
       http.send();
     });
   }