+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+Version 0.1.3 (May 13th 2018)
+-----------------------------
+ * New bind class for data binding.
+ * New file class for file name and path Utils.
+ * New img and sound class for images and audio loading.
+
 Version 0.1.2 (April 29th 2018)
 -----------------------------
  * New ajax class with promises.